+++ src/AjgarlagPsrHttpMessageBundle.php
@@ -12,8 +12,23 @@
 
 namespace Ajgarlag\Bundle\PsrHttpMessageBundle;
 
+use Ajgarlag\Bundle\PsrHttpMessageBundle\DependencyInjection\Compiler\RegisterHttpMessageFactoriesPass;
+use Ajgarlag\Bundle\PsrHttpMessageBundle\DependencyInjection\Compiler\RegisterNyholmPsr17FactoriesPass;
+use Ajgarlag\Bundle\PsrHttpMessageBundle\DependencyInjection\Compiler\TagArgumentValueResolverPass;
+use Ajgarlag\Bundle\PsrHttpMessageBundle\DependencyInjection\Compiler\TagViewEventListenerPass;
+use Symfony\Component\DependencyInjection\Compiler\PassConfig;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class AjgarlagPsrHttpMessageBundle extends Bundle
 {
+    public function build(ContainerBuilder $container)
+    {
+        parent::build($container);
+
+        $container->addCompilerPass(new TagArgumentValueResolverPass(), PassConfig::TYPE_BEFORE_OPTIMIZATION, 1);
+        $container->addCompilerPass(new RegisterHttpMessageFactoriesPass());
+        $container->addCompilerPass(new RegisterNyholmPsr17FactoriesPass());
+        $container->addCompilerPass(new TagViewEventListenerPass());
+    }
 }