+++ tests/DependencyInjection/AjgarlagPsrHttpMessageExtensionTest.php
@@ -13,20 +13,22 @@
 namespace Ajgarlag\Bundle\PsrHttpMessageBundle\Tests\DependencyInjection;
 
 use Ajgarlag\Bundle\PsrHttpMessageBundle\DependencyInjection\AjgarlagPsrHttpMessageExtension;
+use Ajgarlag\Bundle\PsrHttpMessageBundle\DependencyInjection\Compiler\RegisterHttpMessageFactoriesPass;
+use Ajgarlag\Bundle\PsrHttpMessageBundle\DependencyInjection\Compiler\RegisterNyholmPsr17FactoriesPass;
+use Ajgarlag\Bundle\PsrHttpMessageBundle\DependencyInjection\Compiler\TagArgumentValueResolverPass;
+use Ajgarlag\Bundle\PsrHttpMessageBundle\Request\ArgumentValueResolver\Psr7ServerRequestResolver;
+use PHPUnit\Framework\TestCase;
 use Symfony\Bridge\PsrHttpMessage\HttpFoundationFactoryInterface;
 use Symfony\Bridge\PsrHttpMessage\HttpMessageFactoryInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
-class AjgarlagPsrHttpMessageExtensionTest extends \PHPUnit\Framework\TestCase
+class AjgarlagPsrHttpMessageExtensionTest extends TestCase
 {
     public function testDefaultConfiguration()
     {
         $container = new ContainerBuilder();
-        $extension = new AjgarlagPsrHttpMessageExtension();
         $config = [];
-
-        $extension->load([$config], $container);
-        $extension->process($container);
+        $this->compileExtension($container, $config);
 
         $this->assertTrue($container->has(HttpMessageFactoryInterface::class));
         $this->assertTrue($container->has(HttpFoundationFactoryInterface::class));
@@ -42,18 +44,25 @@ class AjgarlagPsrHttpMessageExtensionTest extends \PHPUnit\Framework\TestCase
         $this->assertFalse($container->hasAlias('sensio_framework_extra.psr7.listener.response'));
     }
 
+    private function compileExtension(ContainerBuilder $container, array $config): void
+    {
+        $extension = new AjgarlagPsrHttpMessageExtension();
+        $extension->load([$config], $container);
+        $extension->process($container);
+        (new TagArgumentValueResolverPass())->process($container);
+        (new RegisterHttpMessageFactoriesPass())->process($container);
+        (new RegisterNyholmPsr17FactoriesPass())->process($container);
+    }
+
     public function testSensioFrameworkExtraAliasConfiguration()
     {
         $container = new ContainerBuilder();
-        $extension = new AjgarlagPsrHttpMessageExtension();
         $config = [
             'alias_sensio_framework_extra_services' => [
                 'enabled' => true,
             ],
         ];
-
-        $extension->load([$config], $container);
-        $extension->process($container);
+        $this->compileExtension($container, $config);
 
         $this->assertTrue($container->has('ajgarlag_psr_http_message.psr7.http_message_factory'));
         $this->assertTrue($container->has('ajgarlag_psr_http_message.psr7.http_foundation_factory'));
@@ -73,16 +82,13 @@ class AjgarlagPsrHttpMessageExtensionTest extends \PHPUnit\Framework\TestCase
         }
 
         $container = new ContainerBuilder();
-        $extension = new AjgarlagPsrHttpMessageExtension();
         $config = [
             'alias_sensio_framework_extra_services' => [
                 'enabled' => (bool) random_int(0, 1),
             ],
         ];
-
         $container->setParameter('ajgarlag_psr_http_message_sensio_psr7_enabled', true);
-        $extension->load([$config], $container);
-        $extension->process($container);
+        $this->compileExtension($container, $config);
 
         $this->assertTrue($container->has('ajgarlag_psr_http_message.psr7.http_message_factory'));
         $this->assertTrue($container->has('ajgarlag_psr_http_message.psr7.http_foundation_factory'));